test: rewrite checksum parsing tests and fixtures
- drop old checksum detector and normalizer tests; add a consolidated
  test_checksum_parsers suite covering find_checksum_entry,
  parse_checksum_file, parse_all_checksums and detect_hash_type_from_checksum_filename
- add fixtures under tests/fixtures/checksums for real-world data
- update integration tests to reflect new fixture layout and paths
- improve coverage for parsing, hashing, and format detection
