+++ tests/fixtures/checksums/Joplin-3.6.10.AppImage.sha512
@@ -1,0 +1,1 @@
+c3906aaee17988889247d1846aa37a95b896b1cddf3b6e0cf03db902ac6a42fdbd94b56e45dd1b8fad1418bb31ebf6018fc9460f73b5aae2cfbb2d0f5f134225
