+++ tests/fixtures/checksums/QOwnNotes-x86_64.AppImage.sha256sum
@@ -1,0 +1,1 @@
+9f788283ced4b8576c366e961aa590644536d39699123bfef9128fb1cc10d4ed  QOwnNotes-x86_64.AppImage
